Задание 5.15
